adding orders to a different page

+++ src/Components/OrderCard/index.jsx
@@ -1,25 +1,35 @@
 import { XMarkIcon } from "@heroicons/react/24/solid";
 
-const OrderCard = props => {
-  const {id, title,imagesUrl,price,handleDelete} = props
+const OrderCard = (props) => {
+  const { id, title, imagesUrl, price, handleDelete } = props;
 
+  let renderXMarkIcon;
+  if (handleDelete) {
+    renderXMarkIcon = 
+      <XMarkIcon
+        onClick={() => handleDelete(id)}
+        className="h-6 w-6 text-black cursor-pointer"
+      ></XMarkIcon>
+  }
 
   return (
-    <div  className='flex justify-between items-center mb-3'>
-      <div className='flex items-center gap-2'>
-        <figure className='w-20 h-20'>
-          <img className='w-full h-full rounded-lg object-cover' src={imagesUrl?.[0]} alt={title} />
+    <div className="flex justify-between items-center mb-3">
+      <div className="flex items-center gap-2">
+        <figure className="w-20 h-20">
+          <img
+            className="w-full h-full rounded-lg object-cover"
+            src={imagesUrl?.[0]}
+            alt={title}
+          />
         </figure>
-        <p className='text-sm font-light'>{title}</p>
+        <p className="text-sm font-light">{title}</p>
       </div>
-      <div className='flex items-center gap-4'>
-        <p className='text-lg font-medium '>${price}</p>
-        <XMarkIcon 
-        onClick={() => handleDelete(id)}
-        className="h-6 w-6 text-black cursor-pointer"></XMarkIcon>
+      <div className="flex items-center gap-4">
+        <p className="text-lg font-medium ">${price}</p>
+        {renderXMarkIcon}
       </div>
     </div>
   );
 };
 
-export default OrderCard
+export default OrderCard;
